+++ ios/Runner/AppDelegate.swift
@@ -24,5 +24,8 @@ import UIKit
 
   func didInitializeImplicitFlutterEngine(_ engineBridge: FlutterImplicitEngineBridge) {
     GeneratedPluginRegistrant.register(with: engineBridge.pluginRegistry)
+    if let registrar = engineBridge.pluginRegistry.registrar(forPlugin: "OnDeviceLlmPlugin") {
+      OnDeviceLlmPlugin.register(with: registrar)
+    }
   }
 }